Fixed player XP to be read as a 32 bit unsigned integer.

+++ src/Poe/Components/Player.cs
@@ -19,7 +19,7 @@ namespace PoeHUD.Poe.Components
             }
         }
 
-        public long XP => Address != 0 ? M.ReadUInt(Address + 0x48) : 0L;
+        public uint XP => Address != 0 ? M.ReadUInt(Address + 0x48) : 0;
 		public int Strength => Address != 0 ? M.ReadInt(Address + 0x4c) : 0;
 		public int Dexterity => Address != 0 ? M.ReadInt(Address + 0x50) : 0;
 		public int Intelligence => Address != 0 ? M.ReadInt(Address + 0x54) : 0;
